Fix circular import error by moving rate limiting to separate module
- Created api/rate_limiting.py to house limiter and rate limiting functions
- Updated api/main.py to import from new rate_limiting module
- Updated api/routes/v2/activities.py to import from rate_limiting module
- Resolves Render deployment failure due to circular import
